+++ frontend/src/index.js
@@ -1,8 +1,20 @@
 import Main from "./components/Main";
+import { io } from "socket.io-client";
+
+global.socket = io("http://localhost:3000");
 
 function init() {
   const container = document.getElementById("root");
-  new Main(container);
+  const main = new Main(container);
+
+  socket.emit("join");
+
+  socket.on("joined", (data) => {
+    socket.on("dataFlow", (data) => {
+      const enemyData = data;
+      main.updateData(enemyData);
+    });
+  });
 }
 
 init();