+++ server/handlers/changePlayerData.js
@@ -1,7 +1,14 @@
-function changePlayerData(io) {
+const { RoomsManager } = require("../components/RoomsManager");
+
+const manager = new RoomsManager();
+
+function changePlayerData(io, data) {
   const socket = this;
   const gameId = socket.request.session.gameId;
-  io.to(gameId).emit("receiveData");
+  const playerId = socket.request.session.playerId;
+  manager.changePlayerData(gameId, playerId, data);
+  const response = manager.getEnemyData(gameId, playerId);
+  socket.broadcast.to(gameId).emit("dataFlow", response);
 }
 
 module.exports = {