+++ server/handlers/joinRoomHandler.js
@@ -3,6 +3,7 @@ const { obtainFreeRoom } = require("../utils/obtainFreeRoom");
 function joinRoomHandler(io) {
   const socket = this;
   const room = obtainFreeRoom();
+  socket.request.session.playerId = rooms[room.id].inGame;
   rooms[room.id].inGame++;
   socket.join(room.id);
   socket.request.session.gameId = room.id;