+++ server/CHANGELOG.md
@@ -1,10 +1,11 @@
 # 1.0 Express server setup + socket.io
+
 # 1.1 27.05 - INFA XD - joining rooms based on socket.io + session as auth
+
 SERVER
 |
-`- components ( classes )
-|
-`- handlers ( SOCKET.IO handlers 1 by 1 file )
+`- components ( classes ) | `- handlers ( SOCKET.IO handlers 1 by 1 file )
 |
 `- utils
 
+# 1.2 29.05 - Data flow on server, emitting and handling, fully functional moving.